requirements.txt
Flask==2.2.5
Flask_SQLAlchemy==3.0.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0